Defined wic benchmark resources
