+++ FieldOpt/Reservoir/tests/test_resource_wic_welldir.h
@@ -17,8 +17,8 @@
    along with FieldOpt.  If not, see <http://www.gnu.org/licenses/>.
 ******************************************************************************/
 
-#ifndef FIELDOPT_TEST_RESOURCE_WIC_H
-#define FIELDOPT_TEST_RESOURCE_WIC_H
+#ifndef FIELDOPT_TEST_RESOURCE_WIC_WELLDIR_H
+#define FIELDOPT_TEST_RESOURCE_WIC_WELLDIR_H
 
 #include <QStringList>
 #include <QString>
@@ -48,86 +48,17 @@ class WellDir {
   // Variables: folder path
   QString well_data_dir_ = "../wic-benchmark-wells";
   // QString* well_data_dir_ptr = &well_data_dir_; // obsolete
+  bool debug_ = false;
 };
 
-class WIData {
- public:
-  // Constructor
-  WIData(){};
-
-  // Functions
-  void ReadData(QString file_name);
-
-  // Variables:
-  Matrix<int, Dynamic, 4> IJK;
-  Matrix<double,Dynamic,1> WCF;
-  QString name;
-
-  Matrix<int, Dynamic, 4>* IJK_ptr = &IJK;
-  Matrix<double,Dynamic,1>* WCF_ptr = &WCF;
-
- private:
-
-};
-
-void WIData::ReadData(QString file_name) {
-
-    QFile file(file_name);
-    file.open(QIODevice::ReadOnly|QIODevice::Text);
-
-    QTextStream in(&file);
-    QStringList in_fields;
-
-    Matrix<int, 1, 4> temp_IJK;
-    Matrix<int, Dynamic, 4> IJK_stor;
-    std::vector<double> wcf;
-
-    while(!in.atEnd()) {
-
-        QString line = in.readLine();
-
-        if (line.contains("OPEN")) {
-
-            // TODO: is there a more robust way to read line such that
-            // indices are not wrong if a change in spacing, for example?
-            // how to remove the space that the line begins with?
-            in_fields = line.split(QRegExp("\\s+"));
-
-            // Read & store well name from current line
-            name = in_fields[1];
-
-            // Read IJK values from current line
-            temp_IJK << in_fields[2].toInt(), in_fields[3].toInt(),
-                in_fields[4].toInt(), in_fields[5].toInt();
-
-            // Store IJK values
-            Matrix<int, Dynamic, 4> IJK_curr(IJK_stor.rows() + temp_IJK.rows(), 4);
-            IJK_curr << IJK_stor, temp_IJK;
-            IJK_stor = IJK_curr;
-
-            // Store well connection factor values
-            wcf.push_back(in_fields[8].toFloat());
-        };
-    }
-
-    file.close();
-
-    IJK = IJK_stor;
-    WCF = Map<Matrix<double, Dynamic, 1>>(wcf.data(), wcf.size());
-
-//		return WIData;
-
-    //        debug
-    //        for( int ii=0; ii < in_fields.length(); ++ii ) {
-    //            std::cout << ii << ":" << in_fields[ii].toStdString() << std::endl;
-    //        }
-}
 
 QList<QStringList> WellDir::GetWellDir() {
 
     // Check directory exists
     if (!Utilities::FileHandling::DirectoryExists(well_data_dir_))
-        throw std::runtime_error("Well dir " + well_data_dir_.toStdString() + " not found.");
+        throw std::runtime_error("Well dir "
+                                     + well_data_dir_.toStdString()
+                                     + " not found.");
 
     // Make list of well dirs
     QStringList dir_list_ = MakeDirList(well_data_dir_);
@@ -140,21 +71,23 @@ QList<QStringList> WellDir::GetWellDir() {
     well_list_.append(AddFilesToList(dir_list_, QString("*RMS.DATA")));
     well_list_.append(AddFilesToList(dir_list_, QString("*PCG.DATA")));
 
-    // Check lists are OK
-    std::cout << "well_data_path_: " << well_data_dir_.toStdString() << std::endl;
-    std::cout << "size of well_list_: " << well_list_.size() << std::endl;
+    // Debug: check lists are OK
+    if (debug_){
+        std::cout << "well_data_path_: " << well_data_dir_.toStdString() << std::endl;
+        std::cout << "size of well_list_: " << well_list_.size() << std::endl;
 
-    QStringList temp = well_list_[0];
-    QString temp_str = "well dir list";
-    printWellDirList(temp, temp_str);
+        QStringList temp = well_list_[0];
+        QString temp_str = "well dir list";
+        printWellDirList(temp, temp_str);
 
-    temp = well_list_[1];
-    temp_str = "rms_list";
-    printWellDirList(temp, temp_str);
+        temp = well_list_[1];
+        temp_str = "rms_list";
+        printWellDirList(temp, temp_str);
 
-    temp = well_list_[2];
-    temp_str = "pcg_list";
-    printWellDirList(temp, temp_str);
+        temp = well_list_[2];
+        temp_str = "pcg_list";
+        printWellDirList(temp, temp_str);
+    }
 
     return well_list_;
 }
@@ -177,14 +110,17 @@ QStringList WellDir::AddFilesToList(QStringList temp_dir, QString ext){
     QStringList temp_list_;
         foreach (QString dir_name_, temp_dir){
             QDir temp_dir(dir_name_);
-            QStringList file_name_ = temp_dir.entryList(QStringList() << ext,
-                                                        QDir::AllEntries |
-                                                            QDir::Files |
-                                                            QDir::CaseSensitive |
-                                                            QDir::NoDotAndDotDot);
+            QStringList file_name_ = temp_dir.entryList(
+                QStringList() << ext,
+                QDir::AllEntries |
+                    QDir::Files |
+                    QDir::CaseSensitive |
+                    QDir::NoDotAndDotDot);
             // Check if only one file
             if (file_name_.length()>1)
-                throw std::runtime_error("Too many " + ext.toStdString() + " files in well folder!");
+                throw std::runtime_error("Too many "
+                                             + ext.toStdString()
+                                             + " files in well folder!");
 
             QString file_str_ = dir_name_ + "/" + file_name_[0];
             temp_list_.append(file_str_);
@@ -199,6 +135,5 @@ void WellDir::printWellDirList(QStringList temp, QString temp_str){
         std::cout << temp_str.toStdString() << ": " << temp[ii].toStdString() << std::endl;
     }
 }
-
 }
-#endif //FIELDOPT_TEST_RESOURCE_WIC_H
+#endif //FIELDOPT_TEST_RESOURCE_WIC_WELLDIR_H
